chore: adds session helper exports
