--- packages/payload/src/auth/removeExpiredSessions.ts
@@ -1,10 +1,0 @@
-import type { UserSession } from './types.js'
-
-export const removeExpiredSessions = (sessions: UserSession[]) => {
-  const now = new Date()
-
-  return sessions.filter(({ expiresAt }) => {
-    const expiry = expiresAt instanceof Date ? expiresAt : new Date(expiresAt)
-    return expiry > now
-  })
-}
